save two decimal places of precision

+++ supabase/migrations/20260214090000_add_event_header_image_position_y.sql
@@ -1,3 +1,3 @@
 -- Add optional vertical object-position for event header image.
 ALTER TABLE public.events
-ADD COLUMN IF NOT EXISTS header_image_position_y INTEGER;
+ADD COLUMN IF NOT EXISTS header_image_position_y REAL;
